refactor(headless): isolate recommendation bundle
https://coveord.atlassian.net/browse/KIT-2458
