+++ packages/headless/src/app/engine.ts
@@ -17,6 +17,7 @@ import {
   UpdateAnalyticsConfigurationActionCreatorPayload,
   updateBasicConfiguration,
 } from '../features/configuration/configuration-actions';
+import {versionReducer as version} from '../features/debug/version-slice';
 import {SearchParametersState} from '../state/search-app-state';
 import {matchCoveoOrganizationEndpointUrlAnyOrganization} from '../utils/url-utils';
 import {doNotTrack} from '../utils/utils';
@@ -27,7 +28,6 @@ import {instantlyCallableThunkActionMiddleware} from './instantly-callable-middl
 import {LoggerOptions} from './logger';
 import {logActionErrorMiddleware} from './logger-middlewares';
 import {createReducerManager, ReducerManager} from './reducer-manager';
-import {version} from './reducers';
 import {createRenewAccessTokenMiddleware} from './renew-access-token-middleware';
 import {Store, configureStore} from './store';
 import {ThunkExtraArguments} from './thunk-extra-arguments';
