+++ packages/headless/src/controllers/recommendation/headless-recommendation.ts
@@ -2,12 +2,12 @@ import {NumberValue, Schema, StringValue} from '@coveo/bueno';
 import {Result} from '../../api/search/search/result';
 import {configuration} from '../../app/common-reducers';
 import {RecommendationEngine} from '../../app/recommendation-engine/recommendation-engine';
-import {recommendation} from '../../app/recommendation-reducers';
-import {loadPaginationActions} from '../../features';
+import {loadPaginationActions} from '../../features/pagination/pagination-actions-loader';
 import {
   getRecommendations,
   setRecommendationId,
 } from '../../features/recommendation/recommendation-actions';
+import {recommendationReducer as recommendation} from '../../features/recommendation/recommendation-slice';
 import {loadReducerError} from '../../utils/errors';
 import {validateOptions} from '../../utils/validate-payload';
 import {ErrorPayload} from '../controller/error-payload';
