+++ packages/headless/src/features/fields/fields-actions-loader.ts
@@ -2,7 +2,7 @@ import {AsyncThunkAction, PayloadAction} from '@reduxjs/toolkit';
 import {FieldDescription} from '../../api/search/fields/fields-response';
 import {AsyncThunkSearchOptions} from '../../api/search/search-api-client';
 import {CoreEngine} from '../../app/engine';
-import {fields} from '../../app/reducers';
+import {fieldsReducer as fields} from '../../features/fields/fields-slice';
 import {ConfigurationSection} from '../../state/state-sections';
 import {
   registerFieldsToInclude,
